Fix to set hold current properly

+++ src/lib/axis/motor/stepDir/tmcStepper/StepperSPI.cpp
@@ -142,7 +142,7 @@ int StepDirTmcSPI::modeMicrostepSlewing() {
 
 void StepDirTmcSPI::modeDecayTracking() {
   setDecayMode(settings.decay);
-  driver->rms_current(settings.currentRun*0.7071F, settings.currentHold/settings.currentRun);
+  driver->rms_current(settings.currentRun*0.7071F, (float)settings.currentHold/(float)settings.currentRun);
 }
 
 void StepDirTmcSPI::modeDecaySlewing() {
